Upgrade starlark-go (20240123142251-f86470692795)
Upgrade the Go starlark package to latest one. This involves a rewrite
of the test_configure.py since the starlark resolver configuration is
evaluated part of the exec. This is because using global config vars is
deprecated now. This is not replacing the usage of it yet as it would
involve breaking change of the API.